pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "loguru>=0.7.3",
    "openai>=2.20.0",
    "python-dotenv>=1.2.1",
]
